✨✅: allow github.com/satori/go.uuid identifiers
Add support for UUID as identifier for objects in generic client, also
improved how identifier types are handled, making it easier to add new
ones.
